[macros] Find platform from resolvedExecutable then from target dill.
`resolvedExecutable` is always running on the host, so this makes much
more sense than starting from the target which may be a different
platform. Still try from the target dill if that doesn't work: if the
running tool is a third party tool using the CFE as a library then
that might be the best we have to go on.

R=jakemac@google.com, johnniwinther@google.com

Change-Id: If73dc49e8ee4eb2bc729d9713fa50ddd4ccbccf8
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/359062
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>
Auto-Submit: Morgan :) <davidmorgan@google.com>
Commit-Queue: Morgan :) <davidmorgan@google.com>