[analysis_server] Fix many LSP tests that produce diagnostics
I'm working towards having LSP tests fail by default if error diagnostics are produced. This updates a bunch of tests that currently do, but are not required to.

In many cases these are minor bugs in the test code and non-null variables without values. Some are where we create multiple files after the server is started (which is not atomic) which are generally updated to just create the files before starting up the server instead (as long as the test was not specifically testing open files).

Change-Id: Ia3adfd413d22b2d3d04d6aba9453b919e0aa46dc
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/359224
Commit-Queue: Oleh Prypin <oprypin@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>