Clarify risk and multiple channel CLs in cherry-pick request template.
Clarify the same CP request can be used for both beta and stable and
make more room for doing so, as well as including a whole list of
changelists to be cherry-picked for complicated cases. Clarify distinct
gerrit CLs are needed for beta and stable.

Replace the informationless risk field that everyone says is low with a
mandatory question to explain the actual risk involved.

Fixes: b/329611671
Change-Id: Ibfb81d89722c237e4b8a531bcac2d475798369b7
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/357603
Commit-Queue: Jonas Termansen <sortie@google.com>
Reviewed-by: Kevin Chisholm <kevinjchisholm@google.com>