[analysis_server] Ensure server notifications are sent for newly-opened files whose contents are unmodified
Some time ago I implemented an optimisation to skip analysis when a file is opened in the LSP server if the contents exactly matched what was last analyzed from disk (since it would just produce the same results).

Some notifications (like Outline) are only sent for open files and are triggered by analysis results. Skipping analysis means that some notifications would not be sent when you first open a file (because no new analysis result was computed).

This became quite noticable with the new CodeLenses because there are other CodeLenses (built into the Dart-Code extension) that were relying on Outline notifications and would "time out" after 5s, and this had the effect of causing the new CodeLenses to not appear for 5s (because VS Code waits for all CodeLenses). This wasn't obvious before because of the caching VS Code does for CodeLens locations.

Fixes https://github.com/Dart-Code/Dart-Code/issues/5048

Change-Id: I5653c01fe7b224ba693f5b95a8af8e3c3d3da06d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/360041
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>