[analysis_server] Send Flutter Outline notifications for Flutter framework files
This fixes the check to determine whether a file is Flutter to handle things in Flutter itself.

It also improves the optimization for not setting overlays when they're the same as on-disk to handle files from dependencies (not only those that are explicitly analyzed) which might help avoid https://github.com/dart-lang/sdk/issues/48051#issuecomment-1686202738 more (it was the original intention, but perhaps wasn't working as well as I'd believed).

Fixes https://github.com/Dart-Code/Dart-Code/issues/5045

Change-Id: Ibd04ebeb10e9a1e749f2b44d4053f9b467554f4f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/360042
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>