DAS: wrap new integer division in parens
We use precedence of the parent of the previous `toInt()` invocation to
determine if parentheses are necessary.

The one change to the _existing_ test may be surprising. But here's how
it is: The code pre-fix had unnecessary parentheses, and the fix
previously removed them. That is not the fix's job, so we should keep
whatever (unrelated) parentheses were in the input.

Fixes https://github.com/dart-lang/sdk/issues/54606

Change-Id: I045860a663c3c3a55d93038551f64e658feb56d8
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/360183
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Samuel Rawlins <srawlins@google.com>