Update the style of the AST API docs, part 1
I mostly left the documentation for the classes unchanged in this CL. I
intend to improve the class docs in a second CL, but I thought it would
be too much to try to do it all in one.

As it is, I appologize for the size of the diff. I can break it up if
it's too much to look at. Most of it is simple style updates, formatting
changes, and deletion of duplicated docs on the Impl classes,  but there
are some places where I made more substantive changes.

Change-Id: I5e92dbce8d39bd9c0bf1f804cded1b00526eb082
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/360562
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>