Use the budget to control whether overrides are suggested
I believe that this is the right thing to do from a UX perspective, but
this change could potentially cause flakiness in tests. There's support
in the LSP path for setting a bigger budget, but not for legacy based
tests (at least not yet).

Change-Id: I76353fd3e4cb4cf8fb7a55c27a3353415143dc06
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/360740
Reviewed-by: Keerti Parthasarathy <keertip@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>