[dart2wasm] Give names to closures in names section when possible
Currently for `printInt` below:

  void main() {
    void printInt(int i) {
      print(i);
    }

    [1, 2, 3, 4, 5].forEach(printInt);
  }

We generate the name `$main closure at file:///...`.

With this change, when a closure is a function declaration, we use its
name in the generated name. For the example above we now generate:
`$main closure printInt at file:///...`.

Fixes #55279.

Change-Id: Ia5e8cb512c24ac33be4920c505b9ef33514b10c2
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/361062
Reviewed-by: Jackson Gardner <jacksongardner@google.com>
Commit-Queue: Ömer Ağacan <omersa@google.com>