[analysis_server] Add sync versions of ErrorOr methods and extensions for records
This is a step towards restricting use of `.error` and `.result` in `ErrorOr` to avoid runtime errors (or at least, the wrong runtime errors when error occur).

The reason for adding `.result` was to reduce excessive nesting when functions required the "result" version of multiple things in a chain. This is now handled better using records - we can do:

```
(foo1, foo2).mapResults((foo1, foo2) => ...);
```

This will execute the function only if both values are results and otherwise return the first error. These extensions exist for 2, 3, and 4 field records.

Additionally, I added Sync() versions of several of the helper methods to avoid using `FutureOr` (something that has caused problems with error handline in the past) and needing to `await` in synchronous code that just wants to map results.

This CL does not migrate _all_ uses of `.result` as it would make the CL much larger. Assuming the approach seems good and we will land this, I will migrate the rest.

Change-Id: Ib3f4c390818a8a26467d27c2472f6435055a6bd7
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/361141
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>