[linter] Fix a multiple doc comment bug for `unintended_html_in_doc_comment`.
The lint would trigger on the first doc comment line of a file where the lint was reported. The length of the lint was also overlooked. This CL fixes both problems and reports the correct length and location of each `<>`.

Related to work done in https://dart-review.googlesource.com/c/sdk/+/359900.

Bug: https://github.com/dart-lang/linter/issues/4793
Change-Id: I284378e8cb31ef67dfda82dd0310ca324b3d561a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/361281
Reviewed-by: Phil Quitslund <pquitslund@google.com>
Commit-Queue: Kallen Tu <kallentu@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>