[CFE] Fix incremental compiler re-adding a removed library on changed entry and file invalidation
When changing the entry and the previous entry was modified (and
invalidated) and we use advanced invalidation we have to add the old
entry up front to provide the reference.
If the old entry isn't actually used though it should then later be
removed, which it is.
If it uses some libraries that aren't used otherwise these will be
removed too.

It is then (before this CL) re-added as we convert the just compiled
libraries to dill libaries.
This means that we now have the old entry, but not the sources in the
uri to sources table, nor necessarily the dependencies as they were
removed too.

This CL fixes it by not converting libraries that were removed because
they were unused.

Another solution we might opt for in the future is to see the change of
entry as a change we don't utilize advanced invaidation on.

Change-Id: Ifd7e0668b3e18d0acf090c8d328616de3c14f84a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/361145
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Jens Johansen <jensj@google.com>