Add documentation for lints in the flutter rule set
The tests don't currently pass because the test framework doesn't allow
the use of the 'flutter' package. I have partial support for enabling
this, and an idea for a better solution.

What I need most is verification that the examples and fixes are
reasonable. I'm not as familiar with the more flutter-specific lints,
so I might be off-base on some of them. Of course, any other comments
are also welcome.

Change-Id: I73c8c0565ba262c88f8e363f59aef8bc749894a9
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/359700
Reviewed-by: Marya Belanger <mbelanger@google.com>
Reviewed-by: Phil Quitslund <pquitslund@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>