[CFE] Explicit creation script should work on part files
On presubmit we only run explicit creation on the changed files,
but if any of the changed files is a part that specified "part of" using
an identifier, and we don't otherwise compile that library, we won't
find it and will likely fail because of compilation errors.

This CL identifies such cases, scans the files in the package such a
file is in to find the library with that name, then replaces the part
file with the found file instead when doing the compilation.

This was especially annoying because updating `messages.yaml` would
update `codes_generated.dart` which specifies
`part of _fe_analyzer_shared.messages.codes;`. This should be fixed with
this CL.

Fixes https://github.com/dart-lang/sdk/issues/55236.

Change-Id: I299d5fc98fb3c9781b3134a2e1eff1055da26902
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/361840
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Jens Johansen <jensj@google.com>