[CFE] Batch compiler is a shell for the incremental compiler
This CL replaces the usage of `generateKernelInternal` with a
utilization of the incremental compiler in the so called
"batch compiler" as used by the cfe-strong-linux bot.

Note that bugs *might* lead to sporadic errors in that for instance
errors from one compile might bleed into another compiler (but depend
on which of the launch compilers actually compile it), which
although annoying, reveals issues with the incremental compiler that
we need to fix.
Currently I believe all such issues to have been fixed, but especially
the rapidly evolving macro stuff might introduce stuff in the future.

Timings of running `python3 tools/test.py -n cfe-strong-linux`:

Before:
-j1
[06:17 | 100% | + 9878 | -   35]
=== 9878 tests passed, 35 failed ===

no -j
[01:46 | 100% | + 9878 | -   35]
=== 9878 tests passed, 35 failed ===

After:
-j1
[04:18 | 100% | + 9878 | -   35]
=== 9878 tests passed, 35 failed ===

no -j
[01:20 | 100% | + 9878 | -   35]
=== 9878 tests passed, 35 failed ===

Change-Id: I359cf4c15e4608d036c910c62d7bfa2628dc09a2
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/361660
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Jens Johansen <jensj@google.com>