[analysis_server] Improve the target range for extension type primary constructors
The target range is used by VS Code to render a preview of the target when hovering over a symbol while holding `ctrl`. For a call to the primary constructor of an extension type, the target is a `ConstructorElement` that has a range that covers only the representation type in parens:

```
extension type ET(int a) {}
                 ^^^^^^^
```

This results in "(int a)" being shown as the code preview which is not useful/expected. This change uses the surrounding declaration in this case instead (which matches the behaviour for default constructors of classes).

Fixes https://github.com/Dart-Code/Dart-Code/issues/5005

Change-Id: I24f0a50faf9059b02da07186213d93e3379f15d3
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/354900
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>