Deprecate AstNode.getProperty and AstNode.setProperty.
These methods are no longer used internally by the analyzer, so they
will be removed in a future release of the analyzer. This will allow a
field to be removed from all AST nodes, which should improve the
memory usage of the analyzer.

Clients that still need the functionality provided by getProperty and
setProperty (the ability to attach their own arbitrary data to AST
nodes) can use the Dart Expando class to achieve the same effect.

Change-Id: I7b4a03e000e88ab94ef276da32235c78428d19f7
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/362320
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>